requirements.txt
#requirements
asgiref==3.10.0
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
mysqlclient==2.2.7
openpyxl==3.1.5
pillow==12.0.0
PyJWT==2.10.1
reportlab==4.4.4
sqlparse==0.5.3